% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{aib_font_families}
\alias{aib_font_families}
\title{All AIB Font Families}
\usage{
aib_font_families()
}
\value{
A named list with elements \code{heading}, \code{body}, and \code{serif}.
}
\description{
Returns a named list of font family names for all roles.
}
