% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aib_palettes.R
\name{aib_pal}
\alias{aib_pal}
\title{Annenberg Institute Color Palette Function}
\usage{
aib_pal()
}
\value{
A function that takes an integer \code{n} and returns \code{n} hex color codes.
}
\description{
Returns a function that generates \code{n} colors from the AIB brand palette.
Intended for use inside ggplot2 scale functions.
}
\examples{
pal <- aib_pal()
pal(3)

}
