% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggaib-package.R
\docType{package}
\name{ggaib-package}
\alias{ggaib}
\alias{ggaib-package}
\title{ggaib: ggplot2 Themes for the Annenberg Institute at Brown University}
\description{
Provides ggplot2 themes and color scales following the visual identity of the Annenberg Institute at Brown University. Includes publication, presentation, map, and data-dense theme variants plus discrete, continuous, and diverging color scales using the Institute's brand palette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrewmcamp.github.io/ggaib/}
  \item \url{https://github.com/andrewmcamp/ggaib}
  \item Report bugs at \url{https://github.com/andrewmcamp/ggaib/issues}
}

}
\author{
\strong{Maintainer}: Andrew Camp \email{andrew_camp@brown.edu}

}
\keyword{internal}
