% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_aib.R
\name{scale_fill_aib}
\alias{scale_fill_aib}
\title{Discrete AIB Fill Scale}
\usage{
scale_fill_aib(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 discrete fill scale.
}
\description{
Maps categorical variables to the Annenberg Institute brand palette
for fill aesthetics.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_aib()

}
