% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_aib.R
\name{scale_fill_aib_div}
\alias{scale_fill_aib_div}
\title{Diverging AIB Fill Scale}
\usage{
scale_fill_aib_div(
  low = "#C00404",
  mid = "white",
  high = "#1B3E6F",
  midpoint = 0,
  ...
)
}
\arguments{
\item{low}{Color for the low end. Default is AIB red.}

\item{mid}{Color for the midpoint. Default is white.}

\item{high}{Color for the high end. Default is AIB navy.}

\item{midpoint}{The data value mapped to the midpoint color. Default is 0.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradient2()}}.}
}
\value{
A ggplot2 diverging fill scale.
}
\description{
Creates a diverging fill gradient. Defaults to red-white-navy.
}
